requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-tracking2==0.5.0
django-user-visit==0.4
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.26.4
user-agents==2.2.0
